requirements.txt
customtkinter
pandas
python-docx
openpyxl